# server/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "localpdf REST API (iLovePDF Clone)",
    "version": "1.0.0",
    "description": "High-performance full-stack PDF manipulation and conversion engine."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/upload": {
      "post": {
        "summary": "Upload files (PDFs, images, office documents)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Files uploaded successfully" }
        }
      }
    },
    "/api/merge": {
      "post": {
        "summary": "Merge multiple PDFs into a single document",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "fileId": { "type": "string" },
                        "pages": { "type": "array", "items": { "type": "integer" } }
                      }
                    }
                  }
                },
                "required": ["files"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Merged PDF generated" }
        }
      }
    },
    "/api/split": {
      "post": {
        "summary": "Split PDF by page ranges or extract pages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "mode": { "type": "string", "enum": ["range", "extract", "splitEveryN"] },
                  "ranges": { "type": "string", "example": "1-3, 5, 7-10" },
                  "pages": { "type": "array", "items": { "type": "integer" } },
                  "n": { "type": "integer", "example": 2 }
                },
                "required": ["fileId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Split PDF or ZIP generated" }
        }
      }
    },
    "/api/compress": {
      "post": {
        "summary": "Compress PDF file with quality profiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "level": { "type": "string", "enum": ["low", "medium", "high"], "default": "medium" }
                },
                "required": ["fileId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Compressed PDF generated with savings stats" }
        }
      }
    },
    "/api/convert": {
      "post": {
        "summary": "Convert between PDF, Word (.docx), Excel (.xlsx), HTML, or Images to PDF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "fileIds": { "type": "array", "items": { "type": "string" } },
                  "to": { "type": "string", "enum": ["word", "excel", "html", "pdf"] }
                },
                "required": ["to"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Converted document generated" }
        }
      }
    },
    "/api/edit": {
      "post": {
        "summary": "Edit PDF: add watermark, text annotations, rotate pages, reorder or delete pages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "rotations": { "type": "object" },
                  "deletePages": { "type": "array", "items": { "type": "integer" } },
                  "reorderPages": { "type": "array", "items": { "type": "integer" } },
                  "watermark": {
                    "type": "object",
                    "properties": {
                      "text": { "type": "string" },
                      "opacity": { "type": "number" },
                      "rotationDegrees": { "type": "number" },
                      "fontSize": { "type": "number" }
                    }
                  }
                },
                "required": ["fileId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Edited PDF generated" }
        }
      }
    },
    "/api/protect/protect": {
      "post": {
        "summary": "Encrypt and password protect a PDF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "userPassword": { "type": "string" }
                },
                "required": ["fileId", "userPassword"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Protected PDF generated" }
        }
      }
    },
    "/api/sign": {
      "post": {
        "summary": "Apply electronic signature to PDF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "signatures": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "signatureDataUrl": { "type": "string" },
                        "x": { "type": "number" },
                        "y": { "type": "number" },
                        "width": { "type": "number" },
                        "height": { "type": "number" }
                      }
                    }
                  }
                },
                "required": ["fileId", "signatures"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Signed PDF generated" }
        }
      }
    },
    "/api/ocr": {
      "post": {
        "summary": "Perform OCR on scanned PDF or image",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileId": { "type": "string" },
                  "language": { "type": "string", "default": "eng" }
                },
                "required": ["fileId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Recognized text and searchable PDF generated" }
        }
      }
    },
    "/api/download/{fileId}": {
      "get": {
        "summary": "Download file by ID",
        "parameters": [
          { "name": "fileId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "File stream returned" },
          "404": { "description": "File not found or expired" }
        }
      }
    }
  }
}